backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart

torch
torchvision
Pillow

sentence-transformers
faiss-cpu

langchain
openai

PyMuPDF

redis

pydantic
pydantic-settings